+++ modelopt/torch/export/unified_export_hf_streaming.py
@@ -33,12 +33,8 @@ import torch
 import torch.nn as nn
 from safetensors.torch import save_file
 
-from .quant_aware_conversion import build_reverse_name_mapper
-from .quant_utils import _postprocess_single_tensor, get_quant_config
-from .registry import ExportContext
-from .unified_export_hf import (
+from .hf_export_prep import (
     _add_mtp_exclusions,
-    _dispatch_export_handler,
     _patch_revert_weight_conversion,
     _prepare_moe_inputs,
     _resolve_export_dtype,
@@ -47,6 +43,10 @@ from .unified_export_hf import (
     _warn_on_unsynced_moe_gate_up,
     requantize_resmooth_fused_llm_layers,
 )
+from .quant_aware_conversion import build_reverse_name_mapper
+from .quant_utils import _postprocess_single_tensor, get_quant_config
+from .registry import ExportContext
+from .unified_export_hf import _dispatch_export_handler
 
 __all__ = ["_export_transformers_checkpoint_streaming"]
