+++ tests/gpu/torch/export/test_fsdp2_export.py
@@ -22,11 +22,9 @@ from _test_utils.torch.misc import minimum_sm
 from torch.distributed._composable.fsdp import fully_shard
 
 import modelopt.torch.quantization as mtq
+from modelopt.torch.export.hf_export_prep import requantize_resmooth_fused_llm_layers
 from modelopt.torch.export.layer_utils import is_quantlinear
-from modelopt.torch.export.unified_export_hf import (
-    _export_quantized_weight,
-    requantize_resmooth_fused_llm_layers,
-)
+from modelopt.torch.export.unified_export_hf import _export_quantized_weight
 from modelopt.torch.quantization.utils import fsdp2_aware_weight_update, patch_fsdp_mp_dtypes
